Set up SQL connection and function to display items on run

+++ schema.sql
@@ -5,7 +5,7 @@
     item_id INTEGER(11) AUTO_INCREMENT NOT NULL,
     product_name VARCHAR(100) NOT NULL,
     department_name VARCHAR(100) NOT NULL,
-    price INTEGER(11) NOT NULL,
+    price DECIMAL(10,2) NOT NULL,
     stock_quantity INTEGER(11) NOT NULL,
     PRIMARY KEY (item_id)
     );
